Update tile placement and fix array
When passing through the array into displayController, array would
be an object. On clicking a tile that has been already clicked, it would
change the tile and would repeat.

Removing the {} from return in gameBoard array fixed having the array
be passed as an object. Added more conditions into placeTile and return
condition to determine if board needs to be changed.